# Improve logging in the chunk store.