@pstibrany 2020-11-18
> I realized we are pretty verbose and it doesn't actually add value to logs one stats by one

Added value is seeing in the trace when exactly was operation in given stage, which gives you additional timing information in the log and span. This is now lost (it could easily be put back).

I'm personally fine with that, but would like to double-check with @bboreham and other chunks users.