@pstibrany 2020-11-18
Another difference is that now we will log that summary message even in the case of error, and it will have incorrect (0) values logged.