@cyriltovena 2020-11-18
> Another difference is that now we will log that summary message even in the case of error, and it will have incorrect (0) values logged.



> Added value is seeing in the trace and log when exactly was operation in given stage, which gives you additional timing information. This is now lost (it could easily be put back).

I'm happy to change and keep the trace log, the output log was mostly my concern.