@bboreham 2020-11-18
Yes there's a little bit of timing information lost.
I don't have a strong memory of using that information, but generally everything may be useful.

If this is expensive, it should be worth investigating why.  E.g. is it in some boxing operation where all the arguments are made into `interface{}` ?  Maybe some special-purpose methods on spanlogger would help that, skip the work if debug logging is off and the current span is not sampled.

(I don't think it is reasonable to complain about the amount of work done to log when debug is turned on - this is by definition not something you would normally do in production)