@cyriltovena 2020-11-18
> I don't have a strong memory of using that information, but generally everything may be useful.

I never used the timing informations, I barely used once to debug the count informations.

I used this log debug at high scale and it takes cpu time as this in the hot path, but also allocations, the problem with gokit is that even if you don't have debug activated it still takes ressources.

I could try to keep the timing but only logging in the span, removing the log entirely, happy to spent some time in this direction if we are ok with that.

Current performance improvement should also be taken into consideration 8~10% improvement, for losing timing is not a bad deal :), specially if you never used it or don't use debug.