@bboreham 2020-11-18
> the parameters are evaluated

I don't see anything more complicated than a `len()`, and your change does the same number of those.

Your 8% is 40 microseconds.  Should we look at how come these routines are called so often that that difference is valuable?
If something here is getting called 1,000 times a second we probably don't want to log anything from it, but move the information to a higher level.