@cyriltovena 2020-11-19
> > the parameters are evaluated
> 
> I don't see anything more complicated than a `len()`, and your change does the same number of those.
> 

the biggest pain was the formatMatcher, then the actual log, which I improved already by computing this only once. You're right len() wasn't the problem, although still think gokit is.

> Your 8% is 40 microseconds. Should we look at how come these routines are called so often that that difference is valuable?

Our new Loki calls the index in the ingesters, which causes a multiplication of index queries, although local with boltdb, this was showing in cpu and memory profile.

> If something here is getting called 1,000 times a second we probably don't want to log anything from it, but move the information to a higher level.

That was my first assumption, I thought we didn't really needed this anymore, that's why I packed everything into a single log.
I could remove debug log too, but  I like to run debug in production and  level.Debug() is also taking ressources even if it doesn't output anything.

There's definitely other part to improve, but this one was a low hanging fruit.

Should I just roll back and remove only the multiple occurence of formatMatcher ?