@bboreham 2020-11-19
I don't know what formatMatcher is.  

Definitely sounds like you want to have a different path where this info isn't logged in detail.
Maybe create a new "spam" log level?