@cyriltovena 2020-11-19
This is loki ingester index queries:

![image](https://user-images.githubusercontent.com/1053421/99668547-e6e2ea80-2a3b-11eb-8ed9-1ea5626f959e.png)


As you can see quite some CPU time in the log of the span and gokit, but also chunk.formatMatcher. Allocated object look pretty much the same.

I'll appreciate consideration if you don't use it, turning off debug is not giving much benefits.