@bboreham 2020-11-19
Looks to me like (a) most of the time goes in the actual logging, not in evaluating parameters.

My sense is you don't want this info logged at all, and I would support a flag that gives you that option.