@cyriltovena 2020-11-19
I think Cortex has enough feature flags, I don't think it's bad choosing what is important to log for the user, I've tried to get best of both world here, log less and keep relevant info.