@bboreham 2020-11-19
I don't mean a user-visible flag, I mean in the code.