@bboreham 2020-11-19
I don't see this as a "best of both worlds" thing, I see that you are degrading the information available in one case, and logging millions of lines you are never going to look at in the other case.  Is that incorrect?